Handle <br> tags in changelogs.
Previously, an HTML changelog entry with a <br> would cause the script to fail
to produce any output, so those changes would be omitted from the commit
message. This affected the commits v114r02, v114r04, and v114r05. Now that the
changes are included, the Git hashes have changed, but there's not much we can
do about that now, sadly.