Change Near's committer identifier.
I used "byuu@byuu.org" when I wrote the bsnes-history-kit back in the day,
because that was how he was known at the time. When I made the first version of
ares-history-kit, Near had only recently given up that name, he didn't intend to
ever return to emulation development, so I kept the name the way it was for
consistency.

I intended to keep the identifier the same this time too, so the new ares
history could be cleanly joined with the history previously imported into higan,
but Near had decisively rejected his old name by this point, and (due to fixing
a bug) the history no longer matches the contents of the higan repository
anyway.

The email address is the one GitHub used when he contributed fixes to higan.
It's not his real email address, it doesn't go anywhere, but it's how he chose
to be represented in Git commits, so I want to honour that.