Remove ParaLLEl-RDP code from snapshots.
ares depends on ParaLLEl-RDP for N64 emulation, but it is maintained separately.
To make it easy to keep up-to-date, ares includes a script to download the
latest version from GitHub, and for official releases Near ran that script so
the code would be included in the release archive.

Having a copy of ParaLLEl-RDP in *some* commits makes it difficult to track
history as the code appears and disappears, so we'll just remove it everywhere
for consistency. Additionally, the ParaLLEl-RDP snapshot would often include a
`.git` directory, and trying to commit a `.git` directory to a Git repository is
a Bad Idea, so let's not do that.