Be smarter about creating .gitignore files.
Previously, a .gitignore file was created for every directory that did not
contain any files, even if it contained subdirectories that contained files.
Now, a .gitignore file will only be created if it is genuinely a leaf in the
filesystem tree.

Also, as part of this change I moved 'trim the trailing slash of directory paths
in .zip files' code; as a result, the parallel-rdp directory itself became
"ares/n64/vulkan/parallel-rdp" (no trailing slash) and thus did not contain the
substring "/parallel-rdp/". Since there's no other paths that mention parallel-
rdp in the codebase, I stopped checking for the slashes.