Always use "a" to indicate the second release with a given version.
If Near missed some minor thing, he would often re-roll the release and not bump
the version number. Additionally, if he made more than one date-based release on
the same day, they would necessarily have the same version identifier.

In this collection of archives, I once used "a" as the suffix for the second
release, I once used "_2", and once I used "b" for the second and "a" for the
first. For consistency (and to simplify the regex that recognises version
numbers in filenames) let's just use "a" consistently.